flappy-kiro/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-requests==0.41b0
gunicorn==21.2.0